Move the crypto selection out of the ed25519 logic
Kept the changes minimal and the selection logic the same to avoid too big refactor of the building pipeline.
Pretty sure we can remove the selection from the runtime tho.
